+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ### API breaking changes
 
-* None.
+* Realm-Xcode6.xcodeproj now only builds using Xcode6-Beta4.
 
 ### Enhancements
 
@@ -11,7 +11,8 @@
 
 ### Bugfixes
 
-* None.
+* Fixed bug related to querying on float properties: `floatProperty = 1.7` now works.
+
 
 0.81.0 Release notes (2014-07-22)
 =============================================================